Support multiple icon themes
You can supply multiple icon themes by setting
freedesktop.util.icon_theme to a table that contains icon themes to
search for the requested icon. Example:

freedesktop.utils.icon_theme = { 'Mist', 'gnome' }

This would first try to find the icon in Mist theme and if not found,
in gnome theme.

The change is backwards-compatible. If freedesktop.util.icon_theme
is a string, this icon theme will be used like before this change.

Signed-off-by: Tuomas Jormola <tj@solitudo.net>