Look for icons in all directories from the specification
According to Icon Theme Specification Version 0.11[1], "apps
should look [for icons] in $HOME/.icons (for backwards compatibility),
in $XDG_DATA_DIRS/icons and in /usr/share/pixmaps (in that order)."

Fix icon size handling. The assignment of icon_sizes to isizes in
the lookup_icon() function, is not necessary and caused the former
table to grow rapidly.

[1] http://standards.freedesktop.org/icon-theme-spec/icon-theme-spec-0.11.html